requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
Pillow>=10.0.0
schedule>=1.2.0

# Research & Intelligence
tavily-python>=0.5.0

# Google Workspace
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0

# Email
resend>=2.0.0

# Content Production — Video
fal-client>=0.13.0
moviepy>=2.0.0
imageio[ffmpeg]>=2.30.0
pyav>=15.0.0

# Content Production — Social Posts / Proposal
playwright>=1.40.0
beautifulsoup4>=4.12.0
